Moving everything to its respective foolders
